+++ CHANGES.md
@@ -1,7 +1,11 @@
-0.2.1.0
+0.3.0.0
 =======
 
+-   Allow changing the underlying parser to any parser that implements the
+    `TokenParsing` class from `parsers`
 -   Add `thriftIDLParser` for standard use cases
+-   Add `Language.Thrift.Parser.Trifecta` with a standard Trifecta-based
+    parser
 
 0.2.0.0
 =======